Pre: Move `createWindow` helper in a separate file, since it's not an event handler
